# core: optimize timers, conn handler and responses for throughput under high concurrency