--- a/web-server-test/tests/web-server/http/xexpr.rkt
+++ b/web-server-test/tests/web-server/http/xexpr.rkt
@@ -50,9 +50,9 @@
  (bytes-sort
   (bytes-append
    #"HTTP/1.1 200 OK\r\nDate: REDACTED GMT\r\nLast-Modified: "
-   (string->bytes/utf-8 (seconds->gmt-string 0))
+   (seconds->gmt-bytes 0)
    #"\r\nServer: Racket\r\nContent-Type: text/html; charset=utf-8\r\nConnection: close\r\n\r\n<a href=\"#\">link</a>"))
- 
+
 
  (write-response tm (response/xexpr '(a ([href "#"]) "link")
                                     #:mime-type #"application/xml"))