@Bogdanp 2020-06-14
I've addressed all the comments. The numbers now look like this:

```
Requests      [total, rate, throughput]         240000, 4000.03, 4000.01
Duration      [total, attack, wait]             1m0s, 1m0s, 265.431µs
Latencies     [min, mean, 50, 90, 95, 99, max]  168.712µs, 261.615µs, 248.794µs, 309.237µs, 354.592µs, 541.572µs, 3.149ms
Bytes In      [total, mean]                     1440000, 6.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:240000
Error Set:
```