@Bogdanp 2020-06-16
I plan to do some more profiling in the coming weeks but I don't know exactly when that will be. In the mean time, I think these can be merged.